requirements.txt
SpeechRecognition == 3.8.1
langdetect == 1.0.9
pyTelegramBotAPI == 3.8.2
python_dotenv == 0.19.0
requests == 2.26.0
telebot_calendar == 1.2
python-dotenv == 0.19.0